Switched from Node js to python
This is a test, I want to see if python will work. I already see a error. The video I am following doesn't use a website to communicate from one user to another, so I will need to send the data that goes to one user to a JS file and that file will need to be accessed right away and update the chat.